lifted state of igredients up to app.js
